changed phi handling to q31 and changed observer chaining
