+++ unimoc.mk
@@ -18,6 +18,7 @@ UNIMOCCPPSRC := ${FREEMASTERCPPSRC} \
                 ${BASEDIR}/src/systems.cpp \
                 ${BASEDIR}/src/values.cpp \
                 ${BASEDIR}/src/management.cpp \
+                ${BASEDIR}/src/control_thread.cpp \
                 ${BASEDIR}/src/as5048b.cpp \
 
 # Required include directories
